Wire up beat names to hierarchy level configuration
