+++ v2/helpers/beats.js
@@ -11,24 +11,30 @@ export function beatOneIsPrologue(sortedBookBeats) {
 const CHAPTER_NAMES = ['Act', 'Chapter', 'Scene']
 const BEAT_NAMES = ['Act', 'Beat', 'Scene']
 
-export function beatName(beats, beat) {
+export function beatName(beats, beat, sortedHierarchyLevels) {
   const depth = tree.depth(beats, beat.id)
-  const names = isSeries(beat) ? BEAT_NAMES : CHAPTER_NAMES
-  return names[depth] || `Level-${depth}`
+  const hierarchyLevel = sortedHierarchyLevels[depth]
+  return (hierarchyLevel && hierarchyLevel.name) || `Level-${depth}`
 }
 
-export function beatTitle(beats, beat, offset) {
+export function beatTitle(beats, beat, sortedHierarchyLevels, offset) {
   return beat.title == 'auto'
-    ? i18n(`${beatName(beats, beat)} {number}`, { number: beat.position + offset + 1 })
+    ? i18n(`${beatName(beats, beat, sortedHierarchyLevels)} {number}`, {
+        number: beat.position + offset + 1,
+      })
     : beat.title
 }
 
-export function editingBeatLabel(beats, beat, offset) {
-  return i18n(`${beatName(beats, beat)} {number} title`, { number: beat.position + offset + 1 })
+export function editingBeatLabel(beats, beat, sortedHierarchyLevels, offset) {
+  return i18n(`${beatName(beats, beat, sortedHierarchyLevels)} {number} title`, {
+    number: beat.position + offset + 1,
+  })
 }
 
-export function beatPositionTitle(beats, beat, offset) {
-  return i18n(`${beatName(beats, beat)} {number}`, { number: beat.position + offset + 1 })
+export function beatPositionTitle(beats, beat, sortedHierarchyLevels, offset) {
+  return i18n(`${beatName(beats, beat, sortedHierarchyLevels)} {number}`, {
+    number: beat.position + offset + 1,
+  })
 }
 
 export function insertBeat(position, beats, newId, bookId) {
