+++ v2/index.js
@@ -41,6 +41,7 @@ import * as placeSelectors from './selectors/places'
 import * as tagSelectors from './selectors/tags'
 import * as uiSelectors from './selectors/ui'
 import * as hierarchySelectors from './selectors/hierarchy'
+import * as hierarchyLevelSelectors from './selectors/hierarchyLevel'
 
 import rootReducer from './reducers/root'
 import mainReducer from './reducers/main'
@@ -99,6 +100,7 @@ const selectors = {
   ...tagSelectors,
   ...uiSelectors,
   ...hierarchySelectors,
+  ...hierarchyLevelSelectors,
 }
 
 const actions = {
