+++ v2/selectors/beats.js
@@ -4,6 +4,7 @@ import { currentTimelineSelector } from './ui'
 import { nextId } from '../helpers/beats'
 import { beatTitle, beatOneIsPrologue } from '../helpers/beats'
 import { findNode, children } from '../reducers/tree'
+import { sortedHierarchyLevels } from './hierarchy'
 
 export const allBeatsSelector = (state) => state.beats
 export const beatIdSelector = (state, beatId) => beatId
@@ -50,10 +51,11 @@ export const makeBeatTitleSelector = () =>
   createSelector(
     beatsByBookSelector,
     beatIdSelector,
+    sortedHierarchyLevels,
     positionOffsetSelector,
-    (beats, beatId, positionOffset) => {
+    (beats, beatId, hierarchyLevels, positionOffset) => {
       const beat = findNode(beats, beatId)
       if (!beat) return ''
-      return beatTitle(beats, beat, positionOffset)
+      return beatTitle(beats, beat, hierarchyLevels, positionOffset)
     }
   )
