+++ v2/selectors/hierarchy.js
@@ -1,7 +1,4 @@
 import { createSelector } from 'reselect'
-import { depth } from '../reducers/tree'
-
-import { beatIdSelector, beatsByBookSelector } from './beats'
 
 export const hierarchyLevels = (state) => state.hierarchyLevels
 
@@ -19,14 +16,5 @@ export const sortedHierarchyLevels = createSelector(
       sortedLevels.push(hierarchyLevels[i])
     }
     return sortedLevels
-  }
-)
-
-export const hierarchyLevelSelector = createSelector(
-  beatsByBookSelector,
-  beatIdSelector,
-  sortedHierarchyLevels,
-  (beats, beatId, hierarchyLevels) => {
-    return hierarchyLevels[depth(beats, beatId)] || hierarchyLevels[hierarchyLevels.length - 1]
   }
 )
